requirements.txt
alembic==1.13.1
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
cachelib==0.13.0
click==8.1.7
Eel==0.16.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
msgspec==0.18.6
pyparsing==3.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.3
